mount: move mount utilities from syscall_wrappers.h into mount_utils.h
